Sorted out a bunch of int vs floats, and got characters positioning from level data

+++ Pacman/Actors/Types/Level.hs
@@ -10,5 +10,5 @@ type LevelItem = LevelItemT PickupType WallDirection
 
 type Level = [[LevelItem]]
 
-levelItemSize :: Int
+levelItemSize :: Float
 levelItemSize = 25
