+++ Pacman/Actors/Types/Scene.hs
@@ -5,8 +5,8 @@ import Pacman.Actors.Types.Ghost
 import Pacman.Actors.Types.Level
 
 data Scene = Scene { elapsedTime :: Float,
-                     width :: Int,
-                     height :: Int,
+                     width :: Float,
+                     height :: Float,
                      pacman :: Pacman,
                      ghosts :: [Ghost],
                      level :: Level
